refactor(server): layered architecture with Route → Service → Repository and Hono Context DI
Restructure the webui server into a clean 3-layer architecture:
- Repository layer: data access (filesystem, SQLite, creative-agent SessionStorage)
- Service layer: business logic with factory-based dependency injection
- Route layer: thin HTTP handlers using Hono Context c.get() for service access

Key improvements:
- Eliminate cross-route dependency (conversations importing from config route)
- Encapsulate in-memory state (currentConfig, providerListCache) in service closures
- Decompose conversations.ts (420 lines) into ConversationService + AgentService + thin route (~80 lines)
- Composition root in index.ts wires all dependencies via factory functions
