+++ apps/webui/src/server/repositories/conversation.repo.ts
@@ -1,0 +1,7 @@
+import { createSessionStorage, type SessionStorage } from "@agentchan/creative-agent";
+
+export function createConversationRepo(projectsDir: string): SessionStorage {
+  return createSessionStorage(projectsDir);
+}
+
+export type ConversationRepo = SessionStorage;
